Add files via upload
As part of my machine learning education and using LaTeX, I derived the normal equation and show that regularized linear regression can be interpreted as linear regression with a Bayesian prior normal distribution over the betas. Please note that theta in the markdown notes corresponds to beta in the included derivations.